Add files via upload
Iris Neural Network
Introduction
Artifical Neural Networks are quickly becoming one of the most popular and widely used mechanisms in Machine Learning and Data Analysis. In the last number of years, numerous libraries and software has been developed to equip programmers with a set of tools for modeling and analysing data in order to recognise patterns and make predictions using large data sets. In today's age of Big Data it is important to try make sense of all of the data we have in society. This could range from social media pattern recognitions from anything to finance and economic trends. The reality is that today we have more data in existence than ever before and it growing at a vast and exponential rate.

Artifical Neural Networks aim to mimic and replicate the neurons of a human brain and using the power of the complex mathematical functions allow us to process and model data in such a way that we can form rational assumptions on a given data set.

Given the sheer amount of data out there it is important to note that data we may analyse is often subject to human error and may not always hold a valid essense of truth. For the purpose of this example we will take a look at the Iris Data Set.

More information on the data set can be found on the link provided above or on the front page README of this repository.

Throughout the notebook we aim to build an Artifical Neural Network capable of making predictions of species of Iris Flowers using Keras - Keras is a high-level neural networks API, written in Python and capable of running on top of Tensorflow.

So without further ado, lets get started!

Inputs and Outputs
Data Investigation and Classification
Before trying to create a model for our Neural Network we first need to investigate our data and determine what will be the inputs and what will be our outputs. The CSV file provided contains 5 columns with:

Sepal Length
Sepal Width
Petal Length
Petal Width
Species
Judging by the fact that we are trying to make predictions we must split our data set into sets of:

Inputs - Numerical data values
Outputs - Classification of Iris Flower species
Now that we have the data set loaded we can extract the data we need into appropriate data sets in preparation for training and testing our Model.
Categorical Classification
Here we are using the Keras utility to_categorical() to allow us to turn our output categories into binary class matrices. This is often refered to as "One-Hot" encoding. This is for use with categorical_crossentropy and classification of our species (setosa, versicolor and virginica).

Each Species will be represented as a binary class matrix.

Setosa [1 0 0]
Versicolor [0 1 0]
Virginica [0 0 1]

Creating a Model
Below we can see an example of a how a Neural Network can be visualized. Every Neural Network is made up of these three main consituents.

Input Layer
$x$ number of Hidden Layers
Output Layer
Keras Models
Keras offers a very useful and high level API to handle creation of Neural Networks. The Keras Sequential Model is defined as a linear stack of layers. This is perfect for what we need to create an Artificial Neural Network consisting of Input, Output and Hidden nodes. We define our Model and add the layers to it.

We are trying to create a model that will look somewhat similar to below:

Activation Functions
An Activation Function in a Neural Network defines the output of a given node given its input or set of inputs. Above we applying two activation functions in separate layers.

Sigmoid
A sigmoid function is a mathematical function having an "S" shaped curve (sigmoid curve). Often, sigmoid function refers to the special case of the logistic function shown in the first figure and defined by the formula:

sigmoid

Below we see a plot of the "S" shaped curved or "Sigmoid Curve".
It's usage in neural network are:

Activation function that transform linear inputs to nonlinear outputs.
Bound output to between 0 and 1 so that it can be interpreted as a probability.
Make computation easier than arbitrary activation functions.
Softmax
Softmax regression (or multinomial logistic regression) is a generalization of logistic regression to the case where we want to handle multiple classes.

Here are using Softmax to allow us to let our data flow through the hidden layers and essentially end up as one of our defined classes:

Setosa
Versicolor
Virginica
Configure the Model for training and fit the training data
We configure the Model using the compile() function defined in the Keras Model API. We define an Optimizer, a Loss function and an additional metric - accuracy.

So before we can use our Model for we must first train it. Using the training data subset which we extracted before we can now fit it to our Model.

The goal here is for the Optimizer to essentially minimize the Loss.

We fit the model passing our inputs and our expected outputs and train it across 100 "Epochs" or training cycles. On each iteration we improve the improve the accuracy and miniize the loss.
Making predictions using the Model
To make predictions using our Model we must first prepare the input data to be what the model expects. Here we use a couple of Numpy functions such as around() and expand_dims() to prepare the input data for prediction.

We can then pass get our prediction as a String value from outputs_vals which defined earlier in the Notebook.
Saving and Loading the Model
Keras offers a very simplistic way to save and load your model.

We can easily reload the model in another script using model = load_model("path_to_model.h5")